move ImageViewer related classes to own package
