+++ app/src/main/java/com/derek_s/hubble_gallery/internal/di/ActivityComponent.java
@@ -4,7 +4,7 @@ import com.derek_s.hubble_gallery.base.ActBase;
 import com.derek_s.hubble_gallery.base.FragBase;
 import com.derek_s.hubble_gallery.detailspage.DetailsActivity;
 import com.derek_s.hubble_gallery.home.ActHome;
-import com.derek_s.hubble_gallery.ui.activities.ActImageViewer;
+import com.derek_s.hubble_gallery.imageviewer.ImageViewerActivity;
 import com.derek_s.hubble_gallery.ui.activities.ActWelcome;
 
 import dagger.Component;
@@ -23,7 +23,7 @@ public interface ActivityComponent {
 
   void inject(ActHome app);
 
-  void inject(ActImageViewer app);
+  void inject(ImageViewerActivity app);
 
   void inject(ActWelcome app);
