+++ app/src/main/res-details/layout/act_image_viewer.xml
@@ -3,7 +3,7 @@
              android:layout_width="match_parent"
              android:layout_height="match_parent"
              android:background="@android:color/black"
-             tools:context=".ui.activities.ActImageViewer">
+             tools:context=".imageviewer.ImageViewerActivity">
 
   <com.derek_s.hubble_gallery.ui.widgets.TouchImageView
     android:id="@+id/iv_fullscreen"
